server/requirements.txt
asgiref==3.8.1 ; python_version >= "3.11" and python_version < "4.0"
django-cors-headers==4.7.0 ; python_version >= "3.11" and python_version < "4.0"
django-extensions==3.2.3 ; python_version >= "3.11" and python_version < "4.0"
django-filter==25.1 ; python_version >= "3.11" and python_version < "4.0"
django-rest-framework==0.1.0 ; python_version >= "3.11" and python_version < "4.0"
django==5.1.6 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.11" and python_version < "4.0"
environs==14.1.1 ; python_version >= "3.11" and python_version < "4.0"
marshmallow==3.26.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
psycopg==3.2.5 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "3.13"
tzdata==2025.1 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
